Add files via upload
OBJECTIVE:
How we can build a machine learning system that can predict whether the news is fake or real,
5 Rows  & 20800 Columns,
The objective of this sentiment analysis project is to build a machine learning model that can classify news articles as either real or fake based on their textual content. The project involves preprocessing the data, converting text into numerical features using TF-IDF vectorization, training a Logistic Regression model, and evaluating its performance. Additionally, the project aims to provide visualizations to gain insights into the distribution of real and fake news articles in the dataset.

PRE-PROCESSING:
*Remove missing values and add “NA”
*Text Cleaning (stemming function):
    replace non-alphabetical characters        with spaces.
    converts all text to lowercase to ensure consistency.
     the text into words, creating a list of words.
*It reduces words to their root forms (e.g., "running" to "run") and removes common stopwords (e.g., "the", "and").
*The TF-IDF vectorization step converts the cleaned and stemmed text data into a numerical format that can be used as input for machine learning algorithms.

CONCLUSION:
From this dataset, we can predict whether the news Real or Fake 
Model is performing very well on the test set
In conclusion, this project demonstrates the successful application of machine learning techniques for classifying news articles based on their content. Such models have the potential to aid in the identification of fake news and contribute to media integrity and information reliability. Further improvements and fine-tuning of the model can enhance its accuracy and usefulness in real-world applications.